Add nonlinear CSTR system creation module to `examples/`
Add the `nonlinear_cstr_model.py` module to enable the creation of
`NonlinearSystem` instances that represent a nonlinear Continuous
Stirred Tank Reactor (CSTR) system. This implementation is based
on the example from the paper "Linear Tracking MPC for Nonlinear
Systems—Part II: The Data-Driven Case" by J. Berberich et al.

This addition facilitates the simulation of this system and
enables the reproduction of the paper's example.