Fix typos in docstrings, error messages, and verbose output (#33)
* Fix wrong step printed in closed-loop simulations

The previous step variable printed as a verbose output was the
n-mpc step, which remained constant for n-mpc steps when n-mpc was
greater than 1. Now, the step variable printed is the actual control
step.

* Fix docstring typo and remove comment in LTI controller

- Rename `y_r_param` to `y_s_param` in the docstring of the
  `set_input_output_setpoints` method, as it was a typo.
- Remove leftover comment.

* Fix wrong error message in nonlinear controller

Replace `self.y_s.shape` with `self.y_r.shape` in the `ValueError`
message of the `set_output_setpoint` method, as it was a typo.