Improve documentation and integrate Sphinx with GitHub Pages (#39)
* Use IEEE citation and links for papers in `README.md`

Replace arXiv links and BibTex entries with IEEE versions for
referenced papers to ensure proper attribution.

* Update docstrings to correctly render with Sphinx

- Update function and class docstrings in several modules.
- Add missing module-level docstrings.

This change ensures the project is correctly rendered in
documentation pages built with Sphinx, and improves the overall
documentation presentation.

* Improve readability and structure of `README.md`

- Standardize letter casing for "nonlinear" terminology.
- Fix minor punctuation and formatting inconsistencies.
- Improve structure of the "Controller Creation" section in
  "Code Structure".

* Add Sphinx documentation

* Convert README links to absolute URLs on release for PyPI

Add a step to the "build" job in `release.yml` to convert source
file links in `README.md` from relative paths to absolute GitHub
URLs. This ensures the links point to their corresponding files on
GitHub when rendered on PyPI.

* Add workflow for building & deploying docs to GitHub Pages